Make notify_pm() only decide on sending notification
In order to better abstract and handle the difference between messages
and actions, notify_pm() is changed to only determine whether a private
message or action should send a notification to the user, and leave
sending the notification to the calling method.  This allows PM's and
PA's to have different notification text formats, while still
centralizing the logic to decide when to send them.

TODO: Make notify_pm() actually use logic instead of blindly saying yes.