Implemented custom expression evaluation options
This option allows customization of the boolean logic used to
determine how conditional values are used to filter notifications for
channel messages.  It evaluates as a full boolean logic expression,
including the use of sub-expressions.  The default value of "all" will
bypass this evaluation and simply require all conditions to be true.

The expression consists of space-separated tokens in the following grammar:

* expression = expression operator expression | "(" expression ")" | value
* operator = "and" | "or"
* value = "true" | "false" | condition
* condition = <any condition option>

Also implemented EVAL command for testing purposes.