Implement save/load for user options
In preparation for renaming the project and supparting multiple
notification services, there needs to be a simple method for migrating
user options to the new module name.  Using the built-in dictionary
read/write mechanism, migrating to a new module name would be as simple
as issuing `/msg *notifo save foo` and `/msg *notify load foo`.