Add support for versioning
When building znc-push with the makefile, it will modify the
PUSHVERSION define with version info from `git describe`, build the
module, and then reset PUSHVERSION back to "dev".  When building
directly from znc-buildmod, this will ensure the version is always just
"dev".

A matching command -- `version` -- has been added to the module which
will simply echo the value of PUSHVERSION at the time of compilation.